requirements.txt
Django==1.4.3
django-tastypie==0.9.11
uWSGI==1.4.4
South==0.7.6
MySQL-python==1.2.3
django-openid==0.3a1
django_openid_auth==0.4
iso8601==0.1.4